Ensure dimensions are correctly written when creating debug PBMs.

+++ jbig2dec/jbig2_generic.c
@@ -383,7 +383,9 @@ jbig2_decode_generic_template0(Jbig2Ctx *ctx,
     (void) params;
 
 #ifdef OUTPUT_PBM
-    printf("P4\n%d %d\n", GBW, GBH);
+    printf("P4\n%u %u\n",
+        (uint32_t) GBW,
+        (uint32_t) GBH);
 #endif
 
     if (GBW <= 0)
@@ -596,7 +598,9 @@ jbig2_decode_generic_template1(Jbig2Ctx *ctx,
     (void) params;
 
 #ifdef OUTPUT_PBM
-    printf("P4\n%d %d\n", GBW, GBH);
+    printf("P4\n%u %u\n",
+        (uint32_t) GBW,
+        (uint32_t) GBH);
 #endif
 
     if (GBW <= 0)
@@ -730,7 +734,9 @@ jbig2_decode_generic_template2(Jbig2Ctx *ctx,
     (void) params;
 
 #ifdef OUTPUT_PBM
-    printf("P4\n%d %d\n", GBW, GBH);
+    printf("P4\n%u %u\n",
+        (uint32_t) GBW,
+        (uint32_t) GBH);
 #endif
 
     if (GBW <= 0)
@@ -796,7 +802,9 @@ jbig2_decode_generic_template3(Jbig2Ctx *ctx,
     (void) params;
 
 #ifdef OUTPUT_PBM
-    printf("P4\n%d %d\n", GBW, GBH);
+    printf("P4\n%u %u\n",
+        (uint32_t) GBW,
+        (uint32_t) GBH);
 #endif
 
     if (GBW <= 0)
